Fixes underflow issue with energy calculations
*This fixes an underflow issue with various energy calculations which could
result in acos(arg) resulting in an unchecked NaN. Preliminary tests suggest
this may fix the energy bug associated with regrowing a bead in a short chain.
More tests are needed to confirm, however previous calculations which gave
issues in the first 10 cycles no go all the way to 1000 cycles with no energy
problems.
   The problem here amounted to values for the argument that were almost =-1.0E0_dp.
Setting them equal to -1.0E0_dp did not resolve the issue due to the issue being
inherent in the binary representation of the floating point. In this case the
code should return pi. This sort of error can be detected in the future by using
-fpe0 -traceback as compiler flags. It should be noted that these can result in
false positives.

* Also adjusts an error statement so that it does not print obscenities and instead
gives a useful debugging statement. This debugging statement is also related to
the same underflow.