Adds bondlength regrowth parameters, fixes bugs with bending energy, and bond length distributions
This fix resloves several issues.
1) Previously an error could result in energy calculations when computing the energy of a straight chain.
In energy_intramolecular.F90 and in moves_CBMC.F90 this was fixed by adding if then statments that catch
the cause of this bug and fix the energy calculation. Previously we would get an unhandled acos calcultion
generating a NaN with out warning.

2) added a feature that allows for the selection of maximum and minimum values for a bond lenght regrowth
when doing CBMC with flexible bonds. New variable descriptions were added to the manual along with methodology
used to select the values. Previously, these values were hard coded when could result in a very incorrect bond
length distribution. Additonally the test suite was updated so that these values are read in.

NOTE: THIS UPDATE WILL MEAN YOU WILL HAVE TO CHANGE YOUR OLD TOPMON.INP FILES

3) Fixed a major bug with the generation and selection of bond lengths in the bond length subroutine. Currently,
the code does a two tiered test for generation of bond lengths. First we select a bond length from the distribution
P(L) = L^2 dL. The code for generating this was previously incorrect. Once this value is selected we test it against
the boltzmann weight with a random number. If it fails at this stage it returns to the first stage of bond length
selection. Additionally, comments were added to the code to clarify what is happening.

4) Added a section in the manual describing how to use the specific atom translation portion of the code.