Replace intel-only intrinsic function (norm2(x)) with sqrt(dot_product(x,x))
The function norm2 used in moves_cbmc is intrinsic for ifort (intel compilers only).
Thus norm2(x) is replaced by sqrt(dot_product(x,x)) in all cases.

In fact, the sqrt(dot_product(x,x)) has already been used to calculate the norm in
other parts of the code.

Along the way, as sqrt() is used in calculating the norm, sqrt() is used three-less
times per call of this subroutine, which is expected to result in significant speed up.