Fixes and updates the virial calculations
This commit fixes the old virial coefficient calcualtion code
which has fallen into disrepair and no longer works. It has been
entirely recoded to take advantage of the energy subroutine meaning
that any potential form and/or method developed in the code should
more or less work with the new virial code. I have not yet implemented
quantum corrections or the ability to compute for multiple temperatures
at one time.

The function of the current code is rather simple. Every imv cycles
the snapshot of each of the two molecule's configurations are taken.
The intramolecular energy of each molecule is then computed
The second molecule is then fictitiously placed in the same box as the
first molecule separated by the distance starvir. Via trapezoidal integration
the mayer function (exp(-U_inter*beta)-1)*r^2 is then integraded in steps of
stepvir  until the molecules are a distnace of rmin or smaller apart. The value
of that integral then gets multiplied by the boltzmann weight formed entirely by
the intramolecular energy. References to the method can be found in several
of the TraPPE papers.

To validate the code I ran 3 different simulations. 1 for ethane, 1 for octane
and one for TIP4P water (which includes charges). All values for B2(T) fell well
with in the uncertainy of the published values for each. It should be noted that
TIP4P took many more cycles to converge than the simple TraPPE Alkanes.

Additonally, the manual has been updated to explain the operation of the code in
addition to suggesting some ideas behind best practices. I will note for anyone who
runs this in the future that fully flexible models are difficult to converge due to
the comparitively large phase space.