Add description for a possible error exit
If the maximum displacement is larger than a box size,
which can happen if a box volume is decreasing rapidly
relative to iratio, a translation move can be attempted
which is twice the length of the boxlength. Current pbc
conditions do not account for this rare event, and the
final COM will be outside the box, so this confusing
error for this rare event is explained more.