Fix allow_cutoff_failure and kmax for ewald sum
If allow_cutoff_failure = 1, rcut will be decreased to half
of the minimum boxlength in the case that a volume move decreases
the boxlength to less than 2*rcut, or in the case that the initial
rcut is too small. The maximum displacement in
translation has a max of 2*rcut, so it also needs to be updated
in the case that rcut is changed

Also error exit if trying to do a volume move on box1 when lexzeo = T.

allow_cutoff_failure = 1 now does its thing during
initialization if rcut is too small. Previously this
wasnt implemented so you could start with rcut > boxln/2
if allow_cutoff_failure > 0. Also always checks max displ.
for translation if allow_cutoff_translation = 1 during
initialization.

Maximum displacement in translation is updated if the cutoff must
be changed (allow_cutoff_failure=1) to half the boxlength during
a volume move. This is only done if the volume move is accepted.
If the move is not accepted, the cutoff is changed back to the old
value.

Fix kmax for ewald summation, fix allow_cutoff=1

fix kmax regarding what mansi found.